+++ crates/toasty-driver-dynamodb/src/type.rs
@@ -11,11 +11,16 @@ impl TypeExt for stmt::Type {
         match self {
             stmt::Type::Bool => ScalarAttributeType::N,
             stmt::Type::String => ScalarAttributeType::S,
-            stmt::Type::I8 | stmt::Type::I16 | stmt::Type::I32 | stmt::Type::I64 => {
-                ScalarAttributeType::N
-            }
+            stmt::Type::I8
+            | stmt::Type::I16
+            | stmt::Type::I32
+            | stmt::Type::I64
+            | stmt::Type::U8
+            | stmt::Type::U16
+            | stmt::Type::U32
+            | stmt::Type::U64 => ScalarAttributeType::N,
             stmt::Type::Bytes => ScalarAttributeType::B,
-            _ => todo!("to_ddb_type; ty={:#?}", self),
+            _ => panic!("key attribute must be a string, number, or binary"),
         }
     }
 }